+++ src/Model/Score/Score.php
@@ -33,6 +33,8 @@ use OAT\Library\Lti1p3Core\Util\Collection\CollectionInterface;
  */
 class Score implements ScoreInterface
 {
+    public const DATE_FORMAT = 'Y-m-d\TH:i:s.vP';
+
     /** @var string */
     private $userIdentifier;
 
@@ -225,7 +227,7 @@ class Score implements ScoreInterface
                     'scoreGiven' => $this->scoreGiven,
                     'scoreMaximum' => $this->scoreMaximum,
                     'comment' => $this->comment,
-                    'timestamp' => $this->timestamp->format(DateTimeInterface::ATOM),
+                    'timestamp' => $this->timestamp->format(self::DATE_FORMAT),
                 ]
             ),
             static function ($value): bool {