+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+Unreleased
+-----
+
+* Added `milliseconds` to timestamp in `Score` JSON output
+
 1.2.0
 -----